+++ polar_flow_export/models/fitness_data.py
@@ -23,6 +23,11 @@ class FitnessData(CalendarEvent):
     def _store(self, parent_dir: Path, url_opener: Callable[[str, Optional[Dict[str, Any]]], bytes]) -> List[Path]:
         store_dir = self.get_store_dir(parent_dir)
         out_file = store_dir.joinpath(f"{self._get_datetime()}.txt")
-        with open(out_file, 'w') as out_f:
-            out_f.write(str(self.index))
+        if out_file.exists():
+            self._report_export_already_exists(out_file)
+            return []
+        else:
+            self._report_exporting(out_file)
+            with open(out_file, 'w') as out_f:
+                out_f.write(str(self.index))
         return [out_file]
